Extracción de lógica de filtrado de DOI a Servicio independiente
Se extrajo la lógica para determinar si un item u objeto de DSpace es
elegible para generarle DOI a un servicio externo, de tal forma que
pueda utilzarse sencillamente desde cualquier otra clase.
